Crate emerald_kernel_user_link

Modules

clock

file

graphics

keyboard

mouse

power

process

syscalls

Macros

call_syscall

Creates a syscall, the first argument is the syscall number (in RAX), then the arguments are as follows RCX, RDX, RSI, RDI, R8, R9, R10 (7 arguments max)

sys_arg

Get the syscall arguments from the interrupt state, the arguments come from the registers RCX, RDX, RSI, RDI, R8, R9, R10

to_arg_err

verify_args

Constants

FD_STDERR

FD_STDIN

FD_STDOUT